(GH-78) Fix for Showtitlebar can not fill all screen
Enable/disable the system menu items for minimize and max/restore has an effect to the frame state.

So to fix this issue we must do this at the same time as the other call for updating the system menu by the starting window state before the frame state will be updated for the first time.